# Unroll uint scans