@klauspost 2023-03-26
(benchmarks are a bit noisy on my machine, but trend should be clear)